+++ fpm.groovy
@@ -1,6 +1,7 @@
 import com.concur.*;
 
 workflowDoc = '''
+title: FPM
 overview: Use the FPM tool to create packages for various systems.
 additional_resources:
   - name: FPM
@@ -90,7 +91,7 @@ parameters:
     name: targetTypes
     description: Formats to create with the command.
     default: ['rpm']
-example:
+example: |
   branches:
     feature:
       steps: