+++ gradle.groovy
@@ -2,6 +2,7 @@ concurPipeline  = new com.concur.Commands()
 concurUtil      = new com.concur.Util()
 
 workflowDoc = '''
+title: Gradle
 overview: Execute any Gradle task.
 additional_resources:
   - name: Gradle Official
@@ -58,7 +59,7 @@ parameters:
   - type: List
     name: extraArgs
     description: Any additional arguments to apply to the Gradle task.
-example:
+example: |
   branches:
     feature:
       steps: